+++ assets/src/js/admin/admin-settings.js
@@ -108,19 +108,11 @@ jQuery( document ).ready( function( $ ) {
 			$icon_container = $( 'i', $this ),
 			$loader = $( this ).next(),
 			set_notification_status = $( this ).hasClass( 'give-email-notification-enabled' ) ? 'disabled' : 'enabled',
-			notification_id = $( this ).data( 'id' );
+			notification_id = $( this ).data( 'id' ),
+			canEditEmailNotificationStatus = parseInt( $this.data( 'edit' ) );
 
-		// Bailout if admin can not edit notification status setting.
-		if ( ! parseInt( $this.data( 'edit' ) ) ) {
-			// Remove all notice.
-			$( 'div.give-email-notification-status-notice' ).remove();
-
-			// Add notice.
-			$( 'hr.wp-header-end' ).after( '<div class="updated error give-email-notification-status-notice"><p>' + $( this ).closest( '.give-email-notification-status' ).data( 'notice' ) + '</p></div>' );
-
-			// Scroll to notice.
-			$( 'html, body' ).animate( { scrollTop: $( 'div.give-email-notification-status-notice' ).position().top }, 'slow' );
-
+		if( ! canEditEmailNotificationStatus ) {
+			showEmailNotificationStatusIsNotEditableNotice($this);
 			return false;
 		}
 
@@ -297,6 +289,26 @@ jQuery( document ).ready( function( $ ) {
 
 		preview.val( Give.fn.formatCurrency( '123456.12345', formatterArgs, {} ) );
 	} );
+
+	/**
+	 * Show admin notice if email notification status is not editable.
+	 *
+	 * @unreleased
+	 * @param {object} noticeEditButton
+	 * @return {boolean}
+	 */
+	function showEmailNotificationStatusIsNotEditableNotice( noticeEditButton ){
+		$( 'div.give-email-notification-status-notice' ).remove();
+
+		// Add notice.
+		$( '.wp-heading-inline' ).after( `<div class="updated error give-email-notification-status-notice"><p>${noticeEditButton.data( 'notice' )}</p></div>` );
+
+		// Scroll to notice.
+		let noticeContainer = $( 'div.give-email-notification-status-notice' );
+		if( noticeContainer.length ){
+			$( 'html, body' ).animate( { scrollTop: noticeContainer.position().top }, 'slow' );
+		}
+	}
 } );
 
 // Vertical tabs feature.